move ClippingGroup.js
