+++ docs/api/en/objects/ClippingGroup.html
@@ -2,7 +2,7 @@
 <html lang="en">
 	<head>
 		<meta charset="utf-8" />
-		<base href="../../../../" />
+		<base href="../../../" />
 		<script src="page.js"></script>
 		<link type="text/css" rel="stylesheet" href="page.css" />
 	</head>
@@ -15,6 +15,8 @@
 			A special version of the Group object that defines clipping planes for decendant objects.
 			ClippingGroups can be nested, with clipping planes accumulating by type: intersection or union.
 		</p>
+
+		<p>Note: ClippingGroup is only supported with WebGPURenderer.</p>
 
 		<h2>Constructor</h2>
